% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psygold.R
\docType{package}
\name{psygold-package}
\alias{psygold}
\alias{psygold-package}
\title{psygold: LongGold psychTestR listening tests, performance tests, and questionnaires}
\description{
Implementation of a LongGold test battery consisting of listening tests, performance tests, and questionnaires based on the psychTestR library. Demos of the whole set of tests and questionnaires can be accessed at https://shiny.gold-msi.org/longgold_demo.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fmhoeger/psygold}
  \item Report bugs at \url{https://github.com/fmhoeger/psygold/issues}
}

}
\keyword{internal}
