% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{psygold_deps}
\alias{psygold_deps}
\title{List all psygold dependencies}
\usage{
psygold_deps(recursive = FALSE, repos = getOption("repos"))
}
\arguments{
\item{recursive}{If \code{TRUE}, will also list all dependencies of
psygold packages.}

\item{repos}{The repositories to use to check for updates.
Defaults to \code{getOptions("repos")}.}
}
\description{
List all psygold dependencies
}
