% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{psygold_packages}
\alias{psygold_packages}
\title{List all packages in the psygold}
\usage{
psygold_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include psygold in the list?}
}
\description{
List all packages in the psygold
}
\examples{
psygold_packages()
}
