% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{psygold_update}
\alias{psygold_update}
\title{Update psygold packages}
\usage{
psygold_update(recursive = FALSE, repos = getOption("repos"))
}
\arguments{
\item{recursive}{If \code{TRUE}, will also list all dependencies of
psygold packages.}

\item{repos}{The repositories to use to check for updates.
Defaults to \code{getOptions("repos")}.}
}
\description{
This will check to see if all psygold packages (and optionally, their
dependencies) are up-to-date, and will install after an interactive
confirmation.
}
\examples{
\dontrun{
psygold_update()
}
}
